refactor(ui): merge identity screens into the identity module
`src/ui/identities/` and `src/ui/identity/` differed by one letter and split
one domain across two module trees. Everything from the plural tree now sits
flat in `src/ui/identity/`, matching the one-directory-per-UI-domain layout
used by `ui/tokens/` and `ui/wallets/`; each screen keeps its own internal
multi-file grouping.

The plural `mod.rs` contributed no names the singular one already had, so its
constants, `removed_identities_banner`, `get_selected_wallet`, and test module
merge in byte-identical — only module declarations and the header doc changed.

Paths and imports only; no behavior change.
