+++ src/ui/identity/funding_common.rs
