+++ src/ui/identity/register_dpns_name_screen.rs
