+++ src/ui/identity/top_up_identity_screen/by_using_unused_asset_lock.rs
@@ -2,11 +2,11 @@ use crate::app::AppAction;
 use crate::model::fee_estimation::format_credits_as_dash;
 use crate::ui::MessageType;
 use crate::ui::components::message_banner::MessageBanner;
-use crate::ui::identities::funding_common::{
+use crate::ui::identity::funding_common::{
     FundingAssetLockPicker, FundingMethod, actionable_asset_locks, asset_lock_address,
     asset_lock_status_label,
 };
-use crate::ui::identities::top_up_identity_screen::{TopUpIdentityScreen, WalletFundedScreenStep};
+use crate::ui::identity::top_up_identity_screen::{TopUpIdentityScreen, WalletFundedScreenStep};
 use crate::ui::theme::DashColors;
 use egui::{Color32, Frame, Margin, RichText, Ui};
