+++ src/ui/identity/transfer_screen.rs
@@ -15,8 +15,8 @@ use crate::ui::components::left_panel::add_left_panel;
 use crate::ui::components::styled::island_central_panel;
 use crate::ui::components::top_panel::add_top_panel;
 use crate::ui::components::{BannerHandle, MessageBanner, OptionBannerExt, ResultBannerExt};
-use crate::ui::identities::keys::key_info_screen::KeyInfoScreen;
-use crate::ui::identities::keys::keys_screen::KeysScreen;
+use crate::ui::identity::keys::key_info_screen::KeyInfoScreen;
+use crate::ui::identity::keys::keys_screen::KeysScreen;
 use crate::ui::{MessageType, Screen, ScreenLike};
 use dash_sdk::dashcore_rpc::dashcore::Address;
 use dash_sdk::dashcore_rpc::dashcore::address::NetworkUnchecked;
