+++ src/ui/mod.rs
@@ -21,12 +21,12 @@ use crate::ui::dashpay::qr_code_generator::QRCodeGeneratorScreen;
 use crate::ui::dashpay::send_payment::SendPaymentScreen;
 use crate::ui::dashpay::{DashPayScreen, DashPaySubscreen};
 use crate::ui::dpns::dpns_contested_names_screen::DPNSScreen;
-use crate::ui::identities::keys::add_key_screen::AddKeyScreen;
-use crate::ui::identities::keys::key_info_screen::KeyInfoScreen;
-use crate::ui::identities::keys::keys_screen::KeysScreen;
-use crate::ui::identities::top_up_identity_screen::TopUpIdentityScreen;
-use crate::ui::identities::transfer_screen::TransferScreen;
-use crate::ui::identities::withdraw_screen::WithdrawalScreen;
+use crate::ui::identity::keys::add_key_screen::AddKeyScreen;
+use crate::ui::identity::keys::key_info_screen::KeyInfoScreen;
+use crate::ui::identity::keys::keys_screen::KeysScreen;
+use crate::ui::identity::top_up_identity_screen::TopUpIdentityScreen;
+use crate::ui::identity::transfer_screen::TransferScreen;
+use crate::ui::identity::withdraw_screen::WithdrawalScreen;
 use crate::ui::network_chooser_screen::NetworkChooserScreen;
 use crate::ui::tokens::add_token_by_id_screen::AddTokenByIdScreen;
 use crate::ui::tokens::tokens_screen::{IdentityTokenBasicInfo, IdentityTokenInfo};
@@ -51,10 +51,10 @@ use contracts_documents::update_contract_screen::UpdateDataContractScreen;
 use dash_sdk::dpp::prelude::IdentityPublicKey;
 use dash_sdk::platform::Identifier;
 use dpns::dpns_contested_names_screen::DPNSSubscreen;
-use identities::add_existing_identity_screen::AddExistingIdentityScreen;
-use identities::add_new_identity_screen::AddNewIdentityScreen;
-use identities::register_dpns_name_screen::{RegisterDpnsNameScreen, RegisterDpnsNameSource};
 use identity::IdentityHubScreen;
+use identity::add_existing_identity_screen::AddExistingIdentityScreen;
+use identity::add_new_identity_screen::AddNewIdentityScreen;
+use identity::register_dpns_name_screen::{RegisterDpnsNameScreen, RegisterDpnsNameSource};
 use masternodes::MasternodesScreen;
 use std::fmt;
 use std::sync::Arc;
@@ -120,7 +120,6 @@ pub mod contracts_documents;
 pub mod dashpay;
 pub mod dpns;
 pub mod helpers;
-pub mod identities;
 pub mod identity;
 pub mod masternodes;
 pub mod network_chooser_screen;
