+++ src/ui/tokens/update_token_config.rs
@@ -18,9 +18,9 @@ use crate::ui::components::{MessageBanner, OptionBannerExt, ResultBannerExt};
 use crate::ui::helpers::{
     TransactionType, add_key_chooser, check_token_authorization, render_group_action_text,
 };
-use crate::ui::identities::get_selected_wallet;
-use crate::ui::identities::keys::add_key_screen::AddKeyScreen;
-use crate::ui::identities::keys::key_info_screen::KeyInfoScreen;
+use crate::ui::identity::get_selected_wallet;
+use crate::ui::identity::keys::add_key_screen::AddKeyScreen;
+use crate::ui::identity::keys::key_info_screen::KeyInfoScreen;
 use crate::ui::theme::{ComponentStyles, DashColors, ResponseExt};
 use crate::ui::tokens::validate_signing_key;
 use crate::ui::{MessageType, Screen, ScreenLike};
