+++ src/ui/wallets/import_mnemonic_screen.rs
@@ -4,8 +4,8 @@ use crate::model::wallet::single_key::SingleKeyWallet;
 use crate::ui::components::left_panel::add_left_panel;
 use crate::ui::components::styled::island_central_panel;
 use crate::ui::components::top_panel::add_top_panel;
-use crate::ui::identities::add_existing_identity_screen::AddExistingIdentityScreen;
-use crate::ui::identities::add_new_identity_screen::AddNewIdentityScreen;
+use crate::ui::identity::add_existing_identity_screen::AddExistingIdentityScreen;
+use crate::ui::identity::add_new_identity_screen::AddNewIdentityScreen;
 use crate::ui::{RootScreenType, Screen, ScreenLike};
 
 use crate::model::wallet::Wallet;
